requirements.txt
requests
beautifulsoup4
newspaper3k
python-pptx
cohere
python-dotenv